+++ apps-intern/ungdomsytelse-veileder/env.schema.ts
@@ -17,6 +17,7 @@ const localSchema = z.object({
         .transform((val) => (typeof val === 'boolean' ? val : val === 'true')),
     SIF_PUBLIC_TILLAT_TIDLIG_INNMELDING: z.string().optional().default('off'),
     SIF_PUBLIC_FEATURE_FORLENGE_PERIODE: z.string().optional().default('off'),
+    SIF_PUBLIC_FEATURE_SLETT_SLUTTDATO: z.string().optional().default('off'),
     SIF_PUBLIC_FEATURE_SLETT_AKTIV_DELTAKELSE: z.string().optional().default('off'),
     SIF_PUBLIC_UMAMI_NETTSIDE_ID: z.string().optional(),
     SIF_PUBLIC_NAIS_FRONTEND_TELEMETRY_COLLECTOR_URL: z.string().optional(),