+++ apps-intern/ungdomsytelse-veileder/src/forms/slett-sluttdato-form/slettSluttdatoFormMessages.ts
@@ -1,0 +1,6 @@
+const nb = {
+    'slettSluttdato.validation.bekreftSletting.notChecked': 'Du må bekrefte at sluttdatoen skal slettes',
+};
+export const slettSluttdatoMessages = {
+    nb,
+};