+++ apps-intern/ungdomsytelse-veileder/src/pages/deltaker-page/parts/SlettAktivDeltakerInfo.tsx
@@ -1,4 +1,4 @@
-import { BodyLong, Box, Button, Heading, HGrid, Switch, VStack } from '@navikt/ds-react';
+import { BodyLong, Box, Button, Heading, VStack } from '@navikt/ds-react';
 import { useState } from 'react';
 import { Deltaker } from '../../../types/Deltaker';
 import SlettAktivDeltakerModal from '../../../components/slett-aktiv-deltaker-modal/SlettAktivDeltakerModal';
@@ -10,38 +10,27 @@ interface Props {
 }
 
 const SlettAktivDeltakerInfo = ({ deltaker, deltakelse }: Props) => {
-    const [ekspandert, setEkspandert] = useState(false);
     const [visDialog, setVisDialog] = useState(false);
 
     return (
         <>
-            {deltakelse.erSlettet || !deltakelse.søktTidspunkt ? null : (
-                <VStack gap="space-8">
-                    <Switch checked={ekspandert} onChange={(e) => setEkspandert(e.target.checked)} size="small">
-                        Vis unntakshandlinger
-                    </Switch>
-                    {ekspandert ? (
-                        <HGrid gap="space-4" columns={{ sm: 1, md: '1fr 1fr' }}>
-                            <Box background="warning-moderate" padding="space-24" borderRadius="16">
-                                <VStack gap="space-0">
-                                    <Heading level="3" size="xsmall" spacing>
-                                        Registrer slettet deltakelse i ungdomsprogrammet
-                                    </Heading>
-                                    <BodyLong spacing>
-                                        Hvis en deltaker ikke starter opp i programmet likevel, kan du registrere
-                                        slettet deltakelse her.
-                                    </BodyLong>
-                                    <div>
-                                        <Button variant="secondary" size="small" onClick={() => setVisDialog(true)}>
-                                            Åpne skjema
-                                        </Button>
-                                    </div>
-                                </VStack>
-                            </Box>
-                        </HGrid>
-                    ) : null}
+            <Box background="warning-moderate" padding="space-24" borderRadius="16">
+                <VStack gap="space-0">
+                    <Heading level="3" size="xsmall" spacing>
+                        Registrer slettet deltakelse i ungdomsprogrammet
+                    </Heading>
+                    <BodyLong spacing>
+                        Hvis en deltaker ikke starter opp i programmet likevel, kan du registrere slettet deltakelse
+                        her.
+                    </BodyLong>
+                    <div>
+                        <Button variant="secondary" size="small" onClick={() => setVisDialog(true)}>
+                            Åpne skjema
+                        </Button>
+                    </div>
                 </VStack>
-            )}
+            </Box>
+
             {visDialog && (
                 <SlettAktivDeltakerModal
                     deltaker={deltaker}