+++ apps/omsorgspengesoknad-v2/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @navikt/omsorgspengesoknad-v2
 
+## 0.0.12
+
+### Patch Changes
+
+- Updated dependencies
+    - @navikt/k9-brukerdialog-prosessering-api@0.3.18
+    - @navikt/k9-sak-innsyn-api@0.2.20
+    - @sif/api@0.1.1
+    - @sif/soknad-forms@0.1.1
+    - @sif/soknad-ui@0.1.1
+
 ## 0.0.11
 
 ### Patch Changes