+++ packages/sif-api/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## 0.1.1
+
+### Patch Changes
+
+- Updated dependencies
+    - @navikt/ung-deltakelse-opplyser-api-deltaker@0.3.21
+    - @navikt/k9-brukerdialog-prosessering-api@0.3.18
+    - @navikt/ung-brukerdialog-api@0.0.8
+    - @navikt/k9-sak-innsyn-api@0.2.20
+
 ## 0.1.0
 
 ### Minor Changes