+++ packages/sif-soknad-ui/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @sif/soknad-ui
 
+## 0.1.1
+
+### Patch Changes
+
+- @sif/api@0.1.1
+
 ## 0.1.0
 
 ### Minor Changes